refactor(behaviors): Clean up caps_word
Changed behavior_caps_word.c to be conditionally compiled instead of
wrapping the entire contents of the file in an #if.

Renamed macros that still referred to "break" instead of "continuation".

Changed the type of the continuations count field to allow for more than
255 continuation keys.

Reordered a struct to allow for better packing.

Switched to statically initializing the devs[] array instead of using an
index field in each device's config struct and the init function.

Refactored caps_word_keycode_state_changed_listener() and
caps_word_enhance_usage() to allow for more easily changing
the conditions for what breaks a word and whether mods should be applied
in future commits.